// src/vocabulary_word/dto/word-updation.dto.ts
import { createZodDto } from "nestjs-zod";
import z from "zod";

export const WordUpdationSchema = z.object({
    id: z.uuid({ message: "Invalid word id" }),
    collectionId: z.uuid({ message: "Invalid collection id" }),
    slotId: z.uuid().optional(),
    term: z.string({ message: "Term is required" }),
    meaning: z.string({ message: "Meaning is required" }),
    example: z.string().optional(),
    note: z.string().optional(),
})

export class WordUpdationDto extends createZodDto(WordUpdationSchema) { }